set_real_time_clock: Change parameter from uint32 to uint64.
This should have been done along with the time_t change, but I forgot
to check this then.

Technically this breaks ABI against BeOS, but:
  1. BeOS used an int32, so we'd already slightly broken ABI here
  2. Only one thing at HaikuArchives (VMwareAddons) and one recipe at HaikuPorts
     (samba) uses this function at all.

If it turns out some critical BeOS app uses this, then I guess we can enclose
GCC2 guards around it, but since I can't find any evidence of that, I'm
pushing it without them for now.

+++ headers/os/kernel/OS.h
@@ -374,7 +374,7 @@ extern status_t		convert_to_pthread(thread_id thread, pthread_t *_thread);
 /* Time */
 
 extern uint32		real_time_clock(void);
-extern void			set_real_time_clock(uint32 secsSinceJan1st1970);
+extern void			set_real_time_clock(uint64 secsSinceJan1st1970);
 extern bigtime_t	real_time_clock_usecs(void);
 extern bigtime_t	system_time(void);
 						/* time since booting in microseconds */
